requirements.txt
tqdm==4.66.4
scipy==1.10.1
numpy==1.24.3
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
imageio==2.35.1
Pillow==11.1.0
scikit-image==0.21.0
torch==2.2.0
torchvision==0.15.2
torch-fidelity==0.3.0
torchmetrics==1.4.1
timm==1.0.9
einops==0.7.0
clip==0.2.0
triton==2.0.0
safetensors==0.3.3
tokenizers==0.14.1
transformers==4.36.0
accelerate==0.23.0
diffusers==0.26.3
